+++ src/features/user-stats/get-full-range.ts
@@ -3,17 +3,37 @@ import { DayData } from "shared/types"
 
 // TODO: add tests
 // TODO: extract to constants
-const MILLIS_IN_DAY = 1000 * 3600 * 24
+export const MILLIS_IN_DAY = 1000 * 3600 * 24
 
-// NOTE: return every array of every day with meditation stats from first meditation date
+// NOTE: return array of every meditation for every day from first meditation date
+// TODO: refactor function
+// eslint-disable-next-line max-statements
 export const getFullRange = (daysWithMeditations: DayData[]) => {
-  const daysWithMeditationsAndEmptyDays = daysWithMeditations
+  if (daysWithMeditations.length === 0) return []
+
+  const statisticsWithMissedDays = daysWithMeditations
     .reverse()
     .reduce(fillRangeWithMissedDays, [] as DayData[])
 
-  return daysWithMeditationsAndEmptyDays
+  const todayTimestamp = Timestamp.fromMillis(
+    new Date(Date.now()).setHours(0, 0, 0, 0)
+  )
+  const lastDayInRange =
+    statisticsWithMissedDays[statisticsWithMissedDays.length - 1]
+
+  const lastDayEqualToToday = lastDayInRange.timestamp.isEqual(todayTimestamp)
+  if (lastDayEqualToToday) {
+    return statisticsWithMissedDays
+  }
+
+  const dateDiff = getDateDiffInDays(todayTimestamp, lastDayInRange.timestamp)
+  const missedDays = createMissedDays(dateDiff, todayTimestamp, lastDayInRange)
+  const fullRange = statisticsWithMissedDays.concat(missedDays)
+
+  return fullRange
 }
 
+// eslint-disable-next-line max-statements
 function fillRangeWithMissedDays(
   acc: DayData[],
   day: DayData,
@@ -32,7 +52,7 @@ function fillRangeWithMissedDays(
   return !isCurrentDateEqualToExpected ? accWithMissedDays : [...acc, day]
 }
 
-function getNextDay(date: Timestamp) {
+function getNextDayTimestamp(date: Timestamp) {
   return Timestamp.fromMillis(date.toMillis() + MILLIS_IN_DAY)
 }
 
@@ -41,14 +61,24 @@ function countAndPushMissedDays(
   day: DayData,
   expectedDate: Timestamp
 ) {
-  const dateDiffInDays = Math.ceil(
-    (day.timestamp.toMillis() - expectedDate.toMillis()) / MILLIS_IN_DAY
-  )
+  const dateDiffInDays = getDateDiffInDays(day.timestamp, expectedDate)
+  const missedDays = createMissedDays(dateDiffInDays, expectedDate, day)
 
-  const missedDays = Array(dateDiffInDays)
+  const newAcc = acc.concat(missedDays, day)
+
+  return newAcc
+}
+
+function createMissedDays(
+  dateDiffInDays: number,
+  expectedDate: Timestamp,
+  day: DayData
+) {
+  return Array(dateDiffInDays)
     .fill(null)
     .map((_, i) => {
-      const newTimestamp = i > 0 ? getNextDay(expectedDate) : expectedDate
+      const newTimestamp =
+        i > 0 ? getNextDayTimestamp(expectedDate) : expectedDate
       const emptyDay: DayData = {
         ...day,
         meditations: [],
@@ -57,7 +87,15 @@ function countAndPushMissedDays(
       }
       return emptyDay
     })
-  const newAcc = acc.concat(missedDays, day)
+}
 
-  return newAcc
+function getDateDiffInDays(dayTimestamp: Timestamp, expectedDate: Timestamp) {
+  const diffInDays =
+    (dayTimestamp.toMillis() - expectedDate.toMillis()) / MILLIS_IN_DAY
+
+  if (diffInDays < 0) {
+    throw new Error(`Отрицательная разница между датами`)
+  }
+
+  return Math.ceil(diffInDays)
 }