+++ src/features/user-stats/user-stats.styles.ts
@@ -4,3 +4,15 @@ export const Wrapper = styled.div`
   height: 100%;
   width: 100%;
 `
+
+export const StyledStatNumber = styled.span`
+  font-size: 6rem;
+  line-height: 1;
+  color: var(--c-foreground);
+`
+
+export const StyledStat = styled.span`
+  display: flex;
+  flex-direction: column;
+  align-items: flex-start;
+`