+++ src/shared/constants.ts
@@ -1,7 +1,7 @@
 import { Timestamp } from "firebase/firestore"
 import { DayData, RequestStatus, UserStatsData } from "./types"
 
-export const fibonacciNums = [0, 1, 2, 3, 5, 8, 13, 21, 34, 55, 89]
+export const FIBONACCI_NUMS = [0, 1, 2, 3, 5, 8, 13, 21, 34, 55, 89]
 export const FIB_NUMS_FOR_TIMER = [0, 1, 2, 3, 5, 8, 13, 21]
 export const fibonacciMinsToSeconds = [60, 120, 180, 300, 480, 780, 1260]
 export const fibonacciPercents = [1, 2, 3, 6, 9, 15, 24, 38, 62, 100]